common: fix return value of `store_read`
The function `store_read` returns the size of the block that has just
been read. To be consistent with all other functions that are similar to
that, we should return an `ssize_t` from this function to represent the
read block's size. This also fixes a warning about implicit conversion
losing integer precision.

+++ src/common.c
@@ -343,7 +343,7 @@ out:
     return 0;
 }
 
-int store_read(unsigned char *out, size_t outlen, struct store *store, const struct hash *hash)
+ssize_t store_read(unsigned char *out, size_t outlen, struct store *store, const struct hash *hash)
 {
     int fd, shardfd;
     ssize_t len;
